% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_Beautify_UMAP.R
\name{BeautifyUMAP}
\alias{BeautifyUMAP}
\title{A Self-made Median Function}
\usage{
BeautifyUMAP(
  P1,
  FileName = "NO1",
  RVer = Version,
  LegPos = c(1.01, 0.5),
  LegBox = "vertical",
  LegDir = "vertical",
  LegTextSize = 22,
  TH = 0.5,
  TV = 0.5,
  TitleSize = 30,
  XtextSize = 25,
  YtextSize = 25,
  XaThick = 0.9,
  YaThick = 0.9,
  xangle = 0,
  AxisTitleSize = 1.5,
  AspRat = 1,
  SubTitSize = 15
)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
This function allows you to calculate the median from a numeric vector.
}
\examples{
median_function(seq(1:10))

}
\keyword{median}
